% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_header.R
\name{cond_header}
\alias{cond_header}
\title{Conditional headers}
\usage{
cond_header(cond_value, text = " ", hlevel = 1)
}
\arguments{
\item{cond_value}{An instance of \code{\link{cond_val}}}

\item{text}{character Text of header}

\item{hlevel}{num Level of header}
}
\value{
character Conditional Markdown header of the given level
}
\description{
Insert headers conditionally in the document
}
