% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_knit.R
\name{cond_knit}
\alias{cond_knit}
\title{Custom knit function}
\usage{
cond_knit(inputFile, encoding)
}
\arguments{
\item{inputFile}{File that pass to \code{\link[rmarkdown]{render}}}

\item{encoding}{Encoding system pass to \code{\link[rmarkdown]{render}}}
}
\value{
\code{\link[rmarkdown]{render}} instance
}
\description{
Custom knit function
}
