% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_table.R
\name{cond_table}
\alias{cond_table}
\title{Title}
\usage{
cond_table(df, caption = " ")
}
\arguments{
\item{df}{a data.frame}

\item{caption}{A caption for returning table}
}
\value{
A kableExtra table for 'latex' \code{\link{formato}},
a DT table for 'html' \code{\link{formato}} and
a flextable table for 'docx' \code{\link{formato}}.
}
\description{
Title
}
