% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_val.R
\name{cond_val}
\alias{cond_val}
\title{Logical value to determine conditional options}
\usage{
cond_val(exclude_r_in)
}
\arguments{
\item{exclude_r_in}{List of logical values of formats to exclude R chuncks.
This values default to params. May contain:
\itemize{
\item \code{html}
\item \code{latex}
\item \code{docx}
}}
}
\value{
TRUE if \code{\link{formato}} in \code{exclude_r_in}, FALSE otherwise.
}
\description{
Logical value to determine conditional options
}
\examples{
\dontrun{
```
knitr::opts_chunk$set(
echo = cond_val()
)
```
}

}
