% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formato.R
\name{formato}
\alias{formato}
\title{Get the currently knitting format}
\usage{
formato()
}
\value{
A character string representing the current knitting document format. The 'html' value if interactive session.
}
\description{
Get the currently knitting format
}
