% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdx_docx.R
\name{rmdx_docx}
\alias{rmdx_docx}
\title{Rmdx DOCX document format}
\usage{
rmdx_docx(toc = TRUE, highlight = "pygments", ...)
}
\arguments{
\item{toc}{\link{logical}: See \code{\link[rmarkdown]{word_document}}}

\item{highlight}{\link{character}: See \code{\link[rmarkdown]{html_document}} highlight argument}

\item{...}{other arguments passed to \code{\link[officedown]{rdocx_document}}}
}
\value{
An R Markdown output format object to be passed to \code{\link[rmarkdown]{render}}
}
\description{
\lifecycle{experimental}
}
\details{
This format is based in \code{\link[officedown]{rdocx_document}}.
}
\examples{
\dontrun{
``` yaml
---
...
output: Rmdx::rmdx_docx
---
```
}
}
