% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdx_paper.R
\name{rmdx_paper}
\alias{rmdx_paper}
\title{Rmdx journal document (paper) format}
\usage{
rmdx_paper(...)
}
\arguments{
\item{...}{arguments to be passed to \code{\link[rmarkdown]{pdf_document}}}
}
\value{
An R Markdown output format object to be passed to \code{\link[rmarkdown]{render}}
}
\description{
\lifecycle{experimental}
}
\details{
This format is highly based in \code{\link[rticles]{arxiv_article}}.
The main changes can be summarized as:
\enumerate{
\item An alternative abstract (\code{abstract_alt}) is included with its respective
title (\code{aatitle}). Which can be used to write the abstract in another
language.
\item A short title (\code{short-title}) is included and placed at the top of
the document pages.
}
}
\examples{
\dontrun{
---
...
output: Rmdx::rmdx_paper
---
}
}
