% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdx_docx.R
\name{rmdx_docx}
\alias{rmdx_docx}
\title{Rmdx DOCX document format}
\usage{
rmdx_docx(toc = TRUE)
}
\arguments{
\item{toc}{logical See \code{\link[rmarkdown]{word_document}}}
}
\value{
An R Markdown output format object to be passed to \code{\link[rmarkdown]{render}}
}
\description{
\lifecycle{experimental}
}
\details{
This format is based in \code{\link[officedown]{rdocx_document}}.
}
\examples{
\dontrun{
---
...
output: Rmdx::rmdx_docx
---
}
}
