% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdx_html.R
\name{rmdx_html}
\alias{rmdx_html}
\title{Rmdx HTML document format}
\usage{
rmdx_html(
  toc = TRUE,
  number_sections = TRUE,
  highlight = "pygments",
  footer_name = "Daniel E. de la Rosa",
  footer_copy = "left",
  ...
)
}
\arguments{
\item{toc}{logical See \code{\link[rmarkdown]{html_document}}}

\item{number_sections}{logical See \code{\link[rmarkdown]{html_document}}}

\item{highlight}{character See \code{\link[rmarkdown]{html_document}}}

\item{footer_name}{character author name to be included in footer}

\item{footer_copy}{character much be 'left' or 'right'}

\item{...}{arguments to be passed to \code{\link[bookdown]{html_document2}}}
}
\value{
An R Markdown output format object to be passed to \code{\link[rmarkdown]{render}}
}
\description{
\lifecycle{maturing}
}
\details{
This format is based in \code{\link[bookdown]{html_document2}}.
}
\examples{
\dontrun{
---
...
output: Rmdx::rmdx_html
---
}
}
