% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdx_paper.R
\name{rmdx_paper}
\alias{rmdx_paper}
\title{Rmdx journal document (paper) format}
\usage{
rmdx_paper(...)
}
\arguments{
\item{...}{arguments to be passed to \code{\link[rmarkdown]{pdf_document}}}
}
\value{
An R Markdown output format object to be passed to \code{\link[rmarkdown]{render}}
}
\description{
\lifecycle{experimental}
}
\details{
This format is highly based in \code{\link[rticles]{arxiv_article}}.
}
\examples{
\dontrun{
---
...
output: Rmdx::rmdx_paper
---
}
}
